requirements.txt
llama-index
llama-index-core
python-dotenv
chromadb